6c

+++ 0x00-hello_world/6-size.c
@@ -11,8 +11,8 @@ int main(void)
 
 	printf("The size of a char is: %lu.\n", (unsigned long)sizeof(c));
 	printf("The size of an int is: %lu.\n", (unsigned long)sizeof(i));
-	printf("The size of an int is: %lu.\n", (long)sizeof(i));
-	printf("The size of an int is: %llu.\n", (long long)sizeof(i));
+	printf("The size of an int is: %lu.\n", (unsigned long)sizeof(i));
+	printf("The size of an int is: %lu.\n", (unsigned long)sizeof(i));
 	printf("The size of a float is: %lu.\n", (unsigned long)sizeof(i));
 
 	return (0);
